refactor: 중복 코드 component화 혹은 styles에 정리
기능이 비슷한 코드를 component화하여 재사용성을 높이고
재사용성이 높은 styled component들은 style 폴더에 따로 저장하였습니다.
