+++ FE/src/components/LoginModal/LoginModal.js
@@ -1,9 +1,9 @@
 import React from "react";
 import styled from "styled-components";
-import ModalWrapper from "../../styles/ModalWrapper";
-import ModalTitle from "../../styles/ModalTitle";
-import ModalButton from "../../styles/ModalButton";
-import ModalInputBox from "../../styles/ModalInputBox";
+import ModalWrapper from "../../styles/Modal/ModalWrapper";
+import ModalTitle from "../../styles/Modal/ModalTitle";
+import ModalButton from "../../styles/Modal/ModalButton";
+import ModalInputBox from "../../styles/Modal/ModalInputBox";
 import ModalBackground from "../ModalBackground/ModalBackground";
 import kakao from "../../assets/kakao.png";
 import naver from "../../assets/naver.png";
