+++ FE/src/styles/Modal/DiaryModalHeader.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const DiaryModalHeader = styled.div`
+  width: 100%;
+  display: flex;
+  align-items: flex-end;
+  justify-content: space-between;
+`;
+
+export default DiaryModalHeader;
